Add data grid to IssueBrowse component
Also setup IssueItemDashboard and IssueItemProfile components to use
Id parameter from the IssueBrowse component and load from API the
selected issue in the IssueBrowse component. Created API endpoint to
get just one issue.

This adds technical debt as a generic data grid was not finished, but
I'm willing to accept this in the spirit of getting many other parts
of the Maelstrom Platform implemented sooner than later.

Fixes #44.